+++ @geogebra/solver-sdk/src/graspable-math/create-path-map.ts
@@ -233,6 +233,7 @@ function annotate(
     case 'ReversedOpenClosedRange':
     case 'ReversedClosedOpenRange':
     case 'ReversedClosedRange':
+    case 'Void':
       throw new Error('Unsupported type: ' + tree.type);
     default:
       needToImplement(treeType);