+++ methods/src/test/kotlin/server/TestSelectPlansForPolynomial.kt
@@ -1,5 +1,6 @@
 package server
 
+import methods.fallback.FallbackPlans
 import methods.polynomials.PolynomialsPlans
 import org.junit.jupiter.api.Test
 
@@ -10,6 +11,36 @@ class TestSelectPlansForPolynomial {
             "x(x + 1)",
             setOf(
                 PolynomialsPlans.ExpandPolynomialExpressionInOneVariable,
+            ),
+        )
+    }
+
+    @Test
+    fun testSimplifiedLinearExpression() {
+        testSelectPlanApi(
+            "2x + 1",
+            setOf(
+                FallbackPlans.ExpressionIsFullySimplified,
+            ),
+        )
+    }
+
+    @Test
+    fun testSimplifiedMonomial() {
+        testSelectPlanApi(
+            "-5[x ^ 7]",
+            setOf(
+                FallbackPlans.ExpressionIsFullySimplified,
+            ),
+        )
+    }
+
+    @Test
+    fun testIrreducibleQuadratic() {
+        testSelectPlanApi(
+            "[x^2] + 2x + 10",
+            setOf(
+                FallbackPlans.QuadraticIsIrreducible,
             ),
         )
     }